Add the user current position to the DB User type
Add the `CurrentPos` field to the `database.User` type. This field
contains a number of sector in which the user is.

The player position is written to the database while `AppendUserTrack()`
is being performed. At the same time, the `AppendUserTrack()` function
is moved from the `database_service.go` file to the `battle_service.go`.